refactor(layout): use Layout component for consistent padding and app-feeling on all pages
