+++ src/pages/onboarding/steps/NameStep.tsx
@@ -10,7 +10,8 @@ import {
 } from '@ionic/react';
 import { arrowForward } from 'ionicons/icons';
 import FormField from '@/components/Form/FormField';
-import { useSaveUserName } from '@/hooks';
+import { Layout } from '@/components/Layout/Layout';
+import { useSaveUserName } from '@/hooks/queries';
 
 interface NameStepProps {
   initialName: string;
@@ -49,36 +50,40 @@ export const NameStep: React.FC<NameStepProps> = ({
   };
 
   return (
-    <IonCard className="ion-no-margin">
-      <IonCardContent className="ion-padding">
-        <IonRow>
-          <IonCol>
-            <IonText className="ion-text-center">
-              <h4>Willkommen!</h4>
-              <p>Wie heisst du?</p>
-            </IonText>
-          </IonCol>
-        </IonRow>
+    <Layout>
+      <IonCard className="ion-no-margin">
+        <IonCardContent className="ion-padding">
+          <IonRow>
+            <IonCol>
+              <IonText className="ion-text-center">
+                <h4>Willkommen!</h4>
+                <p>Wie heisst du?</p>
+              </IonText>
+            </IonCol>
+          </IonRow>
 
-        <FormField
-          label=""
-          value={userName}
-          onChange={(value) => setUserName(String(value))}
-          placeholder="Namen eingeben"
-        />
+          <FormField
+            label=""
+            value={userName}
+            onChange={(value) => setUserName(String(value))}
+            placeholder="Namen eingeben"
+          />
 
-        <div className="ion-padding-top">
-          <IonButton
-            expand="block"
-            onClick={handleSaveName}
-            size="small"
-            disabled={saveUserNameMutation.isPending}
-          >
-            {saveUserNameMutation.isPending ? 'Wird gespeichert...' : 'Weiter'}
-            <IonIcon slot="end" icon={arrowForward} />
-          </IonButton>
-        </div>
-      </IonCardContent>
-    </IonCard>
+          <div className="ion-padding-top">
+            <IonButton
+              expand="block"
+              onClick={handleSaveName}
+              size="small"
+              disabled={saveUserNameMutation.isPending}
+            >
+              {saveUserNameMutation.isPending
+                ? 'Wird gespeichert...'
+                : 'Weiter'}
+              <IonIcon slot="end" icon={arrowForward} />
+            </IonButton>
+          </div>
+        </IonCardContent>
+      </IonCard>
+    </Layout>
   );
 };
